move get_value into Task class and rename to get_payload
