+++ Makefile
@@ -67,7 +67,7 @@ test: unit-test integration-test
 
 unit-test:
 ifeq ($(REGEX),)
-	pytest -vvv --log-level=INFO -m unit
+	pytest -vvv --log-level=DEBUG -m unit
 else
 	pytest -vvvv --log-level=INFO -s -m unit -k $(REGEX)
 endif
